+++ libexec/scoop-checkup.ps1
@@ -19,6 +19,7 @@ $issues += !(Test-HelpersInstalled)
 $issues += !(Test-Drive)
 $issues += !(Test-Config)
 $issues += !(Test-CompletionRegistered)
+$issues += !(Test-ShovelAdoption)
 
 if ($issues -gt 0) {
     Write-UserMessage -Message '', "Found $issues potential $(pluralize $issues 'problem' 'problems')." -Warning